requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
djangorestframework==3.13.1
pytz==2021.3
screen==1.0.1
sqlparse==0.4.2
xlwt==1.3.0
